fix call to process kill in the controller when joining

+++ Unity Projet Reva/Assets/Resources/Scripts/TCPController.cs
@@ -110,8 +110,7 @@ public class TCPController : MonoBehaviour {
     private void OnDestroy()
     {
         myTCP.closeSocket();
-
-        process.Kill();
+        if(TCPController.isHosting)process.Kill();
     }
 
     public static IEnumerator StartServer()
